# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Austin Rifle Club API",
    "description": "API for Austin Rifle Club member portal, range status, events, and administration.",
    "version": "1.0.0",
    "contact": {
      "email": "info@austinrifleclub.org"
    }
  },
  "servers": [
    {
      "url": "https://austinrifleclub.org/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787/api",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/payments/checkout/membership": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create membership checkout session",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["membershipType"],
                "properties": {
                  "membershipType": {
                    "type": "string",
                    "enum": ["individual", "family", "life"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": { "type": "string" },
                    "url": { "type": "string", "format": "uri" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/payments/checkout/event/{eventId}": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create event registration checkout session",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Checkout session created" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}